2.0.7 rc

+++ entrypoint.sh
@@ -14,12 +14,37 @@ share() {
     return 0
   fi
 
+  idleTimeout=5
+  # check if .share-artifact-* file was already moved by other thread
+  mv ${LOG_DIR}/.share-artifact-$artifactId ${LOG_DIR}/.sharing-artifact-$artifactId >> /dev/null 2>&1
+  if [ $? -ne 0 ]; then
+    echo "waiting for other thread to share $artifactId files"
+    # if we can't move this file -> other thread already moved it
+    # wait until share is completed on other thread by deleting .recording-artifact-$artifactId file
+    waitStartTime=$(date +%s)
+    while [ $((waitStartTime + idleTimeout)) -gt "$(date +%s)" ]; do
+      if [ ! -f ${LOG_DIR}/.recording-artifact-$artifactId ]; then
+        echo "other thread shared artifacts for $artifactId"
+        return 0
+      fi
+      sleep 0.1
+    done
+    echo "timedout waiting for other thead to share artifacts for $artifactId"
+    return 0
+  fi
+
   # unique folder to collect all artifacts for uploading
   mkdir ${LOG_DIR}/${artifactId}
 
   cp ${TASK_LOG} ${LOG_DIR}/${artifactId}/${LOG_FILE}
   # do not move otherwise in global loop we should add extra verification on file presense
   > ${TASK_LOG}
+
+  if [[ -f ${WDA_LOG_FILE} ]]; then
+    echo "Sharing file: ${WDA_LOG_FILE}"
+    cp ${WDA_LOG_FILE} ${LOG_DIR}/${artifactId}/wda.log
+    > ${WDA_LOG_FILE}
+  fi
 
   if [ "${PLATFORM_NAME}" == "ios" ] && [ -f /tmp/${artifactId}.mp4 ]; then
     ls -la /tmp/${artifactId}.mp4
@@ -29,7 +54,6 @@ share() {
 
     # wait until ffmpeg finished normally and file size is greater 48 byte!
     startTime=$(date +%s)
-    idleTimeout=5
     while [ $(( startTime + idleTimeout )) -gt "$(date +%s)" ]; do
       videoFileSize=$(wc -c /tmp/${artifactId}.mp4  | awk '{print $1}')
       #echo videoFileSize: $videoFileSize
@@ -58,7 +82,6 @@ share() {
     pkill -e -f screenrecord
     # wait until screenrecord finished normally
     startTime=$(date +%s)
-    idleTimeout=5
     while [ $(( startTime + idleTimeout )) -gt "$(date +%s)" ]; do
       echo "detecting screenrecord process pid..."
       screenrecordState=`adb shell "pgrep -l screenrecord" | grep -c screenrecord`
@@ -77,7 +100,7 @@ share() {
 
   # share all the rest custom reports from LOG_DIR into artifactId subfolder
   for file in ${LOG_DIR}/*; do
-    if [ -f "$file" ] && [ -s "$file" ] && [ "$file" != "${TASK_LOG}" ]; then
+    if [ -f "$file" ] && [ -s "$file" ] && [ "$file" != "${TASK_LOG}" ] && [ "$file" != "${VIDEO_LOG}" ] && [ "$file" != "${WDA_LOG_FILE}" ]; then
       echo "Sharing file: $file"
       # to avoid extra publishing as launch artifact for driver sessions
       mv $file ${LOG_DIR}/${artifactId}/
@@ -87,7 +110,13 @@ share() {
   # register artifactId info to be able to parse by uploader
   echo "artifactId=$artifactId" > ${LOG_DIR}/.artifact-$artifactId
 
-  # remove lock file when artifacts are shared for uploader
+  # share video log file
+  cp ${VIDEO_LOG} ${LOG_DIR}/${artifactId}/${VIDEO_LOG_FILE}
+  > ${VIDEO_LOG}
+
+  # remove lock file (for other threads) when artifacts are shared for uploader
+  rm -f ${LOG_DIR}/.sharing-artifact-$artifactId
+  # remove lock file (for uploader) when artifacts are shared for uploader
   rm -f ${LOG_DIR}/.recording-artifact-$artifactId
 }
 
@@ -197,6 +226,8 @@ capture_video() {
     /opt/start-capture-artifacts.sh $recordArtifactId &
     # create .recording-artifact-* file, so uploader would know that recorder is still in process
     echo "artifactId=$recordArtifactId" > ${LOG_DIR}/.recording-artifact-$recordArtifactId
+    # create .share-artifact-* file, so share would be performed only once for session
+    touch ${LOG_DIR}/.share-artifact-$recordArtifactId
 
     # from time to time browser container exited before we able to detect finishedSessionId.
     # make sure to have workable functionality on trap finish (abort use-case as well)
